refactor `impl_from_str` and `impl_from_bytes` macros
